requirements.txt
numpy==1.26.2
opencv-python==4.8.1.78
pillow==10.2.0
pygame==2.5.2
